% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_python}
\alias{check_python}
\title{Check Python environment and required modules}
\usage{
check_python()
}
\value{
Invisibly returns a list with python config and module availability
}
\description{
Check Python environment and required modules
}
