% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_utils.R
\name{clean_names}
\alias{clean_names}
\title{Clean column names to consistent snake_case}
\usage{
clean_names(.data)
}
\arguments{
\item{.data}{Data frame}
}
\value{
Data frame with cleaned names
}
\description{
Clean column names to consistent snake_case
}
