% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hpc_utils.R
\name{cleanup_remote_session}
\alias{cleanup_remote_session}
\title{Clean up remote session directory}
\usage{
cleanup_remote_session(hpc_host, hpc_user = NULL, remote_session_dir)
}
\description{
Clean up remote session directory
}
\keyword{internal}
