% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hpc_utils.R
\name{download_trained_model}
\alias{download_trained_model}
\title{Download trained model from HPC to local machine}
\usage{
download_trained_model(
  hpc_host,
  hpc_user = NULL,
  remote_session_dir,
  output_name,
  local_output_dir
)
}
\description{
Download trained model from HPC to local machine
}
\keyword{internal}
