% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_utils.R
\name{enhance_results}
\alias{enhance_results}
\title{Add useful derived metrics that users would calculate anyway}
\usage{
enhance_results(.data)
}
\arguments{
\item{.data}{Data frame with morphological data}
}
\value{
Data frame with additional calculated metrics
}
\description{
Add useful derived metrics that users would calculate anyway
}
