% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hpc_utils.R
\name{generate_bootstrap_script}
\alias{generate_bootstrap_script}
\title{Generate SLURM script content}
\usage{
generate_bootstrap_script(
  remote_session_dir,
  output_name,
  max_iter,
  learning_rate,
  num_classes,
  eval_period,
  checkpoint_period,
  profile
)
}
\description{
Generate SLURM script content
}
\keyword{internal}
