% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{get_population_stats}
\alias{get_population_stats}
\title{Get overall population statistics}
\usage{
get_population_stats(.data)
}
\arguments{
\item{.data}{Data frame with detections}
}
\value{
Named list of population-level statistics
}
\description{
Get overall population statistics
}
