% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hpc_utils.R
\name{setup_remote_directories}
\alias{setup_remote_directories}
\title{Setup remote directory structure on HPC}
\usage{
setup_remote_directories(hpc_host, hpc_user = NULL, hpc_base_dir, output_name)
}
\description{
Setup remote directory structure on HPC
}
\keyword{internal}
