% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hpc_utils.R
\name{submit_slurm_job}
\alias{submit_slurm_job}
\title{Generate and submit SLURM job for training}
\usage{
submit_slurm_job(
  hpc_host,
  hpc_user = NULL,
  remote_session_dir,
  output_name,
  max_iter,
  learning_rate,
  num_classes,
  eval_period,
  checkpoint_period
)
}
\description{
Generate and submit SLURM job for training
}
\keyword{internal}
