% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summarize_by_image}
\alias{summarize_by_image}
\title{Summarize detections by image}
\usage{
summarize_by_image(.data)
}
\arguments{
\item{.data}{Data frame with detections}
}
\value{
Summary tibble with per-image statistics
}
\description{
Summarize detections by image
}
