% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{summarize_dataset}
\alias{summarize_dataset}
\title{Summarize a dataset directory}
\usage{
summarize_dataset(data_dir)
}
\arguments{
\item{data_dir}{Directory containing 'train' and 'val'}
}
\value{
A tibble with counts for train and val
}
\description{
Summarize a dataset directory
}
