% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hpc_utils.R
\name{sync_code_to_hpc}
\alias{sync_code_to_hpc}
\title{Sync training code to HPC}
\usage{
sync_code_to_hpc(hpc_host, hpc_user = NULL, remote_session_dir)
}
\description{
Sync training code to HPC
}
\keyword{internal}
