% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hpc_utils.R
\name{test_ssh_connection}
\alias{test_ssh_connection}
\title{Test SSH connection to HPC}
\usage{
test_ssh_connection(hpc_host = "hpg", hpc_user = NULL)
}
\description{
Test SSH connection to HPC
}
\keyword{internal}
