% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/training.R
\name{train_model_hpc}
\alias{train_model_hpc}
\title{Train model on HPC using SLURM}
\usage{
train_model_hpc(
  data_dir,
  output_name,
  max_iter,
  learning_rate,
  num_classes,
  eval_period,
  checkpoint_period,
  ims_per_batch,
  num_workers,
  freeze_at,
  optimizer,
  weight_decay,
  backbone_multiplier,
  scheduler,
  warmup_iters,
  steps,
  classwise,
  hpc_env,
  gpus,
  hpc_host,
  hpc_user,
  hpc_base_dir,
  local_output_dir
)
}
\description{
Train model on HPC using SLURM
}
\keyword{internal}
