% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/training.R
\name{train_model_local}
\alias{train_model_local}
\title{Train model locally using available hardware}
\usage{
train_model_local(
  data_dir,
  output_name,
  max_iter,
  learning_rate,
  num_classes,
  device,
  eval_period,
  checkpoint_period,
  local_output_dir
)
}
\description{
Train model locally using available hardware
}
\keyword{internal}
