% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hpc.R
\name{hpc_download}
\alias{hpc_download}
\title{Download trained model from HPC}
\usage{
hpc_download(target, remote_base, output_name, local_output_dir)
}
\description{
Download trained model from HPC
}
\keyword{internal}
