% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hpc_utils.R
\name{hpc_sync_and_submit}
\alias{hpc_sync_and_submit}
\title{Upload data/code and submit SLURM job}
\usage{
hpc_sync_and_submit(
  target,
  data_dir,
  hpc_base_dir,
  output_name,
  training_params,
  gpus = 1
)
}
\description{
Upload data/code and submit SLURM job
}
\keyword{internal}
