% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hpc_utils.R
\name{monitor_slurm_job}
\alias{monitor_slurm_job}
\title{Monitor SLURM job status until completion}
\usage{
monitor_slurm_job(
  hpc_host,
  hpc_user = NULL,
  job_id,
  monitor_interval = 30,
  remote_session_dir
)
}
\description{
Monitor SLURM job status until completion
}
\keyword{internal}
